# api/openapi.yaml
openapi: 3.0.3
info:
  title: Otedama Mining Pool API
  description: |
    Professional P2P Mining Pool & DEX Platform API with BTC-only payouts.
    
    ## Features
    - RESTful API for pool statistics and management
    - WebSocket API for real-time mining operations
    - Comprehensive authentication and security
    - Support for 13 cryptocurrencies and 10 mining algorithms
    
    ## Authentication
    All authenticated endpoints require either:
    - Bearer token (JWT) in Authorization header
    - API key in X-API-Key header
  version: 0.7.0
  contact:
    name: Otedama Support
    url: https://github.com/shizukutanaka/Otedama
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3333
    description: Local development server
  - url: http://localhost:8080
    description: Local API server
...
paths:
...
  /api/miner/{minerId}:
    get:
      tags:
        - Mining
      summary: Get miner details
      description: Retrieve detailed information about a specific miner
      operationId: getMiner
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: minerId
          in: path
          required: true
          description: Miner ID
          schema:
            type: string
      responses:
        '200':
          description: Miner details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Miner'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/payout/request:
    post:
      tags:
        - Wallet
      summary: Request payout
      description: Request a manual payout (automatic payouts occur hourly)
      operationId: requestPayout
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - minerId
              properties:
                minerId:
                  type: string
                  description: Miner ID requesting payout
                amount:
                  type: number
                  description: Amount to payout (optional, defaults to full balance)
                  minimum: 0.001
      responses:
        '200':
          description: Payout request successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/dex/swap:
    post:
      tags:
        - DEX
      summary: Execute swap
      description: Swap tokens using the DEX
      operationId: executeSwap
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapRequest'
      responses:
        '200':
          description: Swap executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    Miner:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
          description: Wallet address
        currency:
          type: string
        algorithm:
          type: string
        created:
          type: integer
          format: int64
        lastSeen:
          type: integer
          format: int64
        totalShares:
          type: integer
        validShares:
          type: integer
        balance:
          type: number
          description: Balance in mined currency
        btcBalance:
          type: number
          description: Balance in BTC
        paid:
          type: number
          description: Total paid amount in BTC
        hashrate:
          type: number
        workers:
          type: integer
          description: Number of active workers
...
    PayoutResponse:
      type: object
      properties:
        transactionId:
          type: string
        amount:
          type: number
          description: Payout amount in BTC
        fee:
          type: number
          description: Network fee
        status:
          type: string
          enum: [pending, processing, completed, failed]
        estimatedTime:
          type: string
          description: Estimated completion time
...
    SwapRequest:
      type: object
      required:
        - tokenIn
        - tokenOut
        - amountIn
      properties:
        tokenIn:
          type: string
        tokenOut:
          type: string
        amountIn:
          type: number
        minAmountOut:
          type: number
          description: Minimum acceptable output amount
        deadline:
          type: integer
          format: int64
          description: Transaction deadline timestamp
    SwapResponse:
      type: object
      properties:
        transactionId:
          type: string
        amountIn:
          type: number
        amountOut:
          type: number
        fee:
          type: number
        executionPrice:
          type: number
        priceImpact:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request parameters
            code: BAD_REQUEST
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
            code: UNAUTHORIZED
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access denied
            code: FORBIDDEN
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
            code: NOT_FOUND
...